feat: split user_profile and user_profile_history modules and refactor data-flow boundaries
