+++ lib/features/item/user_profile/model/user_profile_entity.dart
@@ -1,23 +1,3 @@
-import 'package:flutter/material.dart';
-
-class UserTradeSummary {
-  UserTradeSummary({
-    required this.title,
-    required this.price,
-    required this.date,
-    required this.statusLabel,
-    required this.statusColor,
-    this.thumbnailUrl,
-  });
-
-  final String title;
-  final String price;
-  final String date;
-  final String statusLabel;
-  final Color statusColor;
-  final String? thumbnailUrl;
-}
-
 class UserReview {
   UserReview({
     required this.fromUserId,
@@ -41,7 +21,6 @@ class UserProfile {
     required this.rating,
     required this.reviewCount,
     required this.avatarUrl,
-    required this.trades,
     required this.reviews,
   });
 
@@ -50,6 +29,5 @@ class UserProfile {
   final double rating;
   final int reviewCount;
   final String avatarUrl;
-  final List<UserTradeSummary> trades;
   final List<UserReview> reviews;
 }
